// lib/engine/abi.generated.ts
/* eslint-disable */
// This file is generated from game_engine/od_core. Do not edit by hand.
export const ABI = Object.freeze({
  DRAWCMD_SIZE_BYTES: 32,
  DRAWCMD_PROGRAM_RECT: 0,
  DRAWCMD_PROGRAM_TEXT: 1,
  DRAWCMD_PROGRAM_OFFSET: 0,
  DRAWCMD_INSTANCE_OFFSET_OFFSET: 4,
  DRAWCMD_INSTANCE_COUNT_OFFSET: 8,
  DRAWCMD_SCISSOR_X_OFFSET: 12,
  DRAWCMD_SCISSOR_Y_OFFSET: 16,
  DRAWCMD_SCISSOR_W_OFFSET: 20,
  DRAWCMD_SCISSOR_H_OFFSET: 24,
  DRAWCMD_RESERVED_OFFSET: 28,
  RECT_INSTANCE_STRIDE_FLOATS: 8,
  RECT_INSTANCE_STRIDE_BYTES: 32,
  RECT_INSTANCE_POS_OFFSET: 0,
  RECT_INSTANCE_SIZE_OFFSET: 8,
  RECT_INSTANCE_TINT_OFFSET: 16,
  RECT_INSTANCE_ALPHA_OFFSET: 28,
  GLYPH_INSTANCE_STRIDE_FLOATS: 12,
  GLYPH_INSTANCE_STRIDE_BYTES: 48,
  GLYPH_INSTANCE_POS_OFFSET: 0,
  GLYPH_INSTANCE_SIZE_OFFSET: 8,
  GLYPH_INSTANCE_UV_OFFSET: 16,
  GLYPH_INSTANCE_TINT_OFFSET: 32,
  GLYPH_INSTANCE_ALPHA_OFFSET: 44,
  INPUT_SAMPLED_SIZE_BYTES: 32,
  INPUT_SAMPLE_FRAMEBUFFER_W_OFFSET: 0,
  INPUT_SAMPLE_FRAMEBUFFER_H_OFFSET: 4,
  INPUT_SAMPLE_DPR_OFFSET: 8,
  INPUT_SAMPLE_DT_MS_OFFSET: 12,
  INPUT_SAMPLE_WINDOW_FOCUSED_OFFSET: 16,
  INPUT_SAMPLE_POINTER_X_OFFSET: 20,
  INPUT_SAMPLE_POINTER_Y_OFFSET: 24,
  INPUT_SAMPLE_BUTTONS_OFFSET: 28,
  INPUT_QUEUE_HEADER_SIZE_BYTES: 8,
  INPUT_QUEUE_HEADER_COUNT_OFFSET: 0,
  INPUT_QUEUE_HEADER_OVERFLOW_OFFSET: 4,
  INPUT_EVENT_SIZE_BYTES: 8,
  INPUT_EVENT_KIND_OFFSET: 0,
  INPUT_EVENT_MODIFIERS_OFFSET: 1,
  INPUT_EVENT_CODE_OFFSET: 2,
  INPUT_EVENT_VALUE_OFFSET: 4,
  INPUT_QUEUE_CAPACITY: 1024,
  INPUT_ARENA_SAMPLE_OFFSET: 0,
  INPUT_ARENA_QUEUE_OFFSET: 32,
  INPUT_ARENA_EVENTS_OFFSET: 40,
  INPUT_ARENA_SIZE_BYTES: 8232,
  INPUT_KIND_UNKNOWN: 0,
  INPUT_KIND_KEY_DOWN: 1,
  INPUT_KIND_KEY_UP: 2,
  INPUT_KIND_BLUR: 3,
  INPUT_KIND_RESYNC: 4,
  INPUT_MODIFIER_SHIFT: 1,
  INPUT_MODIFIER_CTRL: 2,
  INPUT_MODIFIER_ALT: 4,
  INPUT_MODIFIER_META: 8,
  KEYCODE_UNKNOWN: 0,
  KEYCODE_ENTER: 1,
  KEYCODE_ESCAPE: 2,
  KEYCODE_SPACE: 3,
  KEYCODE_KEYI: 4,
  KEYCODE_KEYJ: 5,
  KEYCODE_KEYK: 6,
  KEYCODE_KEYL: 7,
  KEYCODE_KEYQ: 8,
  KEYCODE_KEYE: 9,
  KEYCODE_KEYS: 10,
  KEYCODE_KEYD: 11,
  KEYCODE_KEYF: 12,
} as const);
export type DrawCmdProgram = typeof ABI.DRAWCMD_PROGRAM_RECT | typeof ABI.DRAWCMD_PROGRAM_TEXT;
export type EventKind = typeof ABI.INPUT_KIND_UNKNOWN | typeof ABI.INPUT_KIND_KEY_DOWN | typeof ABI.INPUT_KIND_KEY_UP | typeof ABI.INPUT_KIND_BLUR | typeof ABI.INPUT_KIND_RESYNC;
export type KeyCode = typeof ABI.KEYCODE_UNKNOWN | typeof ABI.KEYCODE_ENTER | typeof ABI.KEYCODE_ESCAPE | typeof ABI.KEYCODE_SPACE | typeof ABI.KEYCODE_KEYI | typeof ABI.KEYCODE_KEYJ | typeof ABI.KEYCODE_KEYK | typeof ABI.KEYCODE_KEYL | typeof ABI.KEYCODE_KEYQ | typeof ABI.KEYCODE_KEYE | typeof ABI.KEYCODE_KEYS | typeof ABI.KEYCODE_KEYD | typeof ABI.KEYCODE_KEYF;